Fix to session feature. Remove reviewed cards from session after answer. Semi-fix too long answers - add error message and cut message to max length of telegram message. Fix some markdown parsing errors.

+++ spaced-repetition-bot/src/main/java/org/company/spacedrepetitionbot/handler/handlers/callback/strategy/edit_message/learn/answer/BaseAnswerStrategy.java
@@ -36,8 +36,12 @@ public abstract class BaseAnswerStrategy extends BaseEditCallbackStrategy {
         Long cardId = Long.valueOf(getCallbackDataByIndex(callbackQuery.getData(), 1));
         Long sessionId = Long.valueOf(getCallbackDataByIndex(callbackQuery.getData(), 3));
 
-        learningSessionService.updateCardWithAnswer(cardId, getQuality());
-        learningSessionService.removeCardFromSession(sessionId, cardId);
+        Card updatedCard = learningSessionService.updateCardWithAnswer(cardId, getQuality());
+        if (updatedCard.getStatus() != Status.NEW &&
+                updatedCard.getStatus() != Status.LEARNING &&
+                updatedCard.getStatus() != Status.RELEARNING) {
+            learningSessionService.removeCardFromSession(sessionId, cardId);
+        }
 
         super.executeCallbackQuery(callbackQuery);
     }
